requirements.txt
# torch==2.2.0 
# numpy
ninja
dm-tree
tqdm
torchmetrics==1.3.1
